refactor: create connection flow
